Add a texture

+++ src/VertexArray.cpp
@@ -1,6 +1,6 @@
 #include "VertexArray.h"
 
-#include "Renderer.h"
+#include "ErrorHandler.h"
 
 VertexArray::VertexArray(/* args */) {
     GLCall(glGenVertexArrays(1, &m_RendererID));
@@ -21,7 +21,7 @@ void VertexArray::AddBuffer(const VertexBuffer& vb, const VertexBufferLayout& la
         const auto& element = elements[i];
 
         GLCall(glEnableVertexAttribArray(i));
-        GLCall(glVertexAttribPointer(i, element.count, element.type, element.normalized, layout.GetStride(), (const void*)offset));
+        GLCall(glVertexAttribPointer(i, element.count, element.type, element.normalized, layout.GetStride(), (const void*)(size_t) offset));
 
         offset += element.count * VertexBufferElement::GetSizeOfType(element.type);
     }
